requirements.txt
pandas
psycopg2-binary
peewee
Flask
python-dotenv
plotly
dash